Recognizing admin status
